refactor: extract dump_file
